requirements.txt
tensorflow-gpu=1.2
scikit-image
scipy
numpy

